+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.19 branch
 ### 2.19.10
+* `UploaderComponent::save()` method no longer throws an exception if the
+    destination directory is not writable, but sets an error;
 * fixed bug for `WRITABLE_DIRS` and `VENDOR_LINKS` configuration values. They no
     longer override the values set by other plugins.